chore: moved database and router logic into separate pkgs
