+++ internal/pkg/database/database.go
@@ -1,0 +1,27 @@
+package database
+
+import (
+	"log"
+
+	"gorm.io/driver/sqlite"
+	"gorm.io/gorm"
+)
+
+type Database interface {
+}
+
+func NewDB() *gorm.DB {
+	db, err := gorm.Open(sqlite.Open("test.db"), &gorm.Config{})
+	if err != nil {
+		log.Fatalf("failed to connect database: %v", err)
+	}
+
+	return db
+}
+
+func SetupMigration(db *gorm.DB) {
+	err := db.AutoMigrate(&User{})
+	if err != nil {
+		log.Fatal(err)
+	}
+}
